+++ tests/unit/config/test_project.py
@@ -1,0 +1,501 @@
+from copy import deepcopy
+import json
+import os
+import unittest
+import pytest
+
+from unittest import mock
+
+import dbt.config
+from dbt.constants import DEPENDENCIES_FILE_NAME
+import dbt.exceptions
+from dbt.adapters.factory import load_plugin
+from dbt.adapters.contracts.connection import QueryComment, DEFAULT_QUERY_COMMENT
+from dbt.contracts.project import PackageConfig, LocalPackage, GitPackage
+from dbt.node_types import NodeType
+from dbt_common.semver import VersionSpecifier
+
+from dbt.flags import set_from_args
+from dbt.tests.util import safe_set_invocation_context
+
+
+from tests.unit.config import (
+    BaseConfigTest,
+    project_from_config_norender,
+    empty_project_renderer,
+    project_from_config_rendered,
+)
+
+
+class TestProject(BaseConfigTest):
+    def test_defaults(self):
+        project = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir
+        )
+        self.assertEqual(project.project_name, "my_test_project")
+        self.assertEqual(project.version, "0.0.1")
+        self.assertEqual(project.profile_name, "default")
+        self.assertEqual(project.project_root, self.project_dir)
+        self.assertEqual(project.model_paths, ["models"])
+        self.assertEqual(project.macro_paths, ["macros"])
+        self.assertEqual(project.seed_paths, ["seeds"])
+        self.assertEqual(project.test_paths, ["tests"])
+        self.assertEqual(project.analysis_paths, ["analyses"])
+        self.assertEqual(
+            set(project.docs_paths), set(["models", "seeds", "snapshots", "analyses", "macros"])
+        )
+        self.assertEqual(project.asset_paths, [])
+        self.assertEqual(project.target_path, "target")
+        self.assertEqual(project.clean_targets, ["target"])
+        self.assertEqual(project.log_path, "logs")
+        self.assertEqual(project.packages_install_path, "dbt_packages")
+        self.assertEqual(project.quoting, {})
+        self.assertEqual(project.models, {})
+        self.assertEqual(project.on_run_start, [])
+        self.assertEqual(project.on_run_end, [])
+        self.assertEqual(project.seeds, {})
+        self.assertEqual(project.dbt_version, [VersionSpecifier.from_version_string(">=0.0.0")])
+        self.assertEqual(project.packages, PackageConfig(packages=[]))
+        # just make sure str() doesn't crash anything, that's always
+        # embarrassing
+        str(project)
+
+    def test_eq(self):
+        project = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir
+        )
+        other = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir
+        )
+        self.assertEqual(project, other)
+
+    def test_neq(self):
+        project = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir
+        )
+        self.assertNotEqual(project, object())
+
+    def test_implicit_overrides(self):
+        self.default_project_data.update(
+            {
+                "model-paths": ["other-models"],
+            }
+        )
+        project = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir
+        )
+        self.assertEqual(
+            set(project.docs_paths),
+            set(["other-models", "seeds", "snapshots", "analyses", "macros"]),
+        )
+
+    def test_hashed_name(self):
+        project = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir
+        )
+        self.assertEqual(project.hashed_name(), "754cd47eac1d6f50a5f7cd399ec43da4")
+
+    def test_all_overrides(self):
+        # log-path is not tested because it is set exclusively from flags, not cfg
+        self.default_project_data.update(
+            {
+                "model-paths": ["other-models"],
+                "macro-paths": ["other-macros"],
+                "seed-paths": ["other-seeds"],
+                "test-paths": ["other-tests"],
+                "analysis-paths": ["other-analyses"],
+                "docs-paths": ["docs"],
+                "asset-paths": ["other-assets"],
+                "clean-targets": ["another-target"],
+                "packages-install-path": "other-dbt_packages",
+                "quoting": {"identifier": False},
+                "models": {
+                    "pre-hook": ["{{ logging.log_model_start_event() }}"],
+                    "post-hook": ["{{ logging.log_model_end_event() }}"],
+                    "my_test_project": {
+                        "first": {
+                            "enabled": False,
+                            "sub": {
+                                "enabled": True,
+                            },
+                        },
+                        "second": {
+                            "materialized": "table",
+                        },
+                    },
+                    "third_party": {
+                        "third": {
+                            "materialized": "view",
+                        },
+                    },
+                },
+                "on-run-start": [
+                    "{{ logging.log_run_start_event() }}",
+                ],
+                "on-run-end": [
+                    "{{ logging.log_run_end_event() }}",
+                ],
+                "seeds": {
+                    "my_test_project": {
+                        "enabled": True,
+                        "schema": "seed_data",
+                        "post-hook": "grant select on {{ this }} to bi_user",
+                    },
+                },
+                "data_tests": {"my_test_project": {"fail_calc": "sum(failures)"}},
+                "require-dbt-version": ">=0.1.0",
+            }
+        )
+        packages = {
+            "packages": [
+                {
+                    "local": "foo",
+                },
+                {"git": "git@example.com:dbt-labs/dbt-utils.git", "revision": "test-rev"},
+            ],
+        }
+        project = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir, packages=packages
+        )
+        self.assertEqual(project.project_name, "my_test_project")
+        self.assertEqual(project.version, "0.0.1")
+        self.assertEqual(project.profile_name, "default")
+        self.assertEqual(project.model_paths, ["other-models"])
+        self.assertEqual(project.macro_paths, ["other-macros"])
+        self.assertEqual(project.seed_paths, ["other-seeds"])
+        self.assertEqual(project.test_paths, ["other-tests"])
+        self.assertEqual(project.analysis_paths, ["other-analyses"])
+        self.assertEqual(project.docs_paths, ["docs"])
+        self.assertEqual(project.asset_paths, ["other-assets"])
+        self.assertEqual(project.clean_targets, ["another-target"])
+        self.assertEqual(project.packages_install_path, "other-dbt_packages")
+        self.assertEqual(project.quoting, {"identifier": False})
+        self.assertEqual(
+            project.models,
+            {
+                "pre-hook": ["{{ logging.log_model_start_event() }}"],
+                "post-hook": ["{{ logging.log_model_end_event() }}"],
+                "my_test_project": {
+                    "first": {
+                        "enabled": False,
+                        "sub": {
+                            "enabled": True,
+                        },
+                    },
+                    "second": {
+                        "materialized": "table",
+                    },
+                },
+                "third_party": {
+                    "third": {
+                        "materialized": "view",
+                    },
+                },
+            },
+        )
+        self.assertEqual(project.on_run_start, ["{{ logging.log_run_start_event() }}"])
+        self.assertEqual(project.on_run_end, ["{{ logging.log_run_end_event() }}"])
+        self.assertEqual(
+            project.seeds,
+            {
+                "my_test_project": {
+                    "enabled": True,
+                    "schema": "seed_data",
+                    "post-hook": "grant select on {{ this }} to bi_user",
+                },
+            },
+        )
+        self.assertEqual(
+            project.data_tests,
+            {
+                "my_test_project": {"fail_calc": "sum(failures)"},
+            },
+        )
+        self.assertEqual(project.dbt_version, [VersionSpecifier.from_version_string(">=0.1.0")])
+        self.assertEqual(
+            project.packages,
+            PackageConfig(
+                packages=[
+                    LocalPackage(local="foo", unrendered={"local": "foo"}),
+                    GitPackage(
+                        git="git@example.com:dbt-labs/dbt-utils.git",
+                        revision="test-rev",
+                        unrendered={
+                            "git": "git@example.com:dbt-labs/dbt-utils.git",
+                            "revision": "test-rev",
+                        },
+                    ),
+                ]
+            ),
+        )
+        str(project)  # this does the equivalent of project.to_project_config(with_packages=True)
+        json.dumps(project.to_project_config())
+
+    def test_string_run_hooks(self):
+        self.default_project_data.update(
+            {
+                "on-run-start": "{{ logging.log_run_start_event() }}",
+                "on-run-end": "{{ logging.log_run_end_event() }}",
+            }
+        )
+        project = project_from_config_rendered(self.default_project_data)
+        self.assertEqual(project.on_run_start, ["{{ logging.log_run_start_event() }}"])
+        self.assertEqual(project.on_run_end, ["{{ logging.log_run_end_event() }}"])
+
+    def test_invalid_project_name(self):
+        self.default_project_data["name"] = "invalid-project-name"
+        with self.assertRaises(dbt.exceptions.DbtProjectError) as exc:
+            project_from_config_norender(self.default_project_data, project_root=self.project_dir)
+
+        self.assertIn("invalid-project-name", str(exc.exception))
+
+    def test_no_project(self):
+        os.remove(os.path.join(self.project_dir, "dbt_project.yml"))
+        renderer = empty_project_renderer()
+        with self.assertRaises(dbt.exceptions.DbtProjectError) as exc:
+            dbt.config.Project.from_project_root(self.project_dir, renderer)
+
+        self.assertIn("No dbt_project.yml", str(exc.exception))
+
+    def test_invalid_version(self):
+        self.default_project_data["require-dbt-version"] = "hello!"
+        with self.assertRaises(dbt.exceptions.DbtProjectError):
+            project_from_config_norender(self.default_project_data, project_root=self.project_dir)
+
+    def test_unsupported_version(self):
+        self.default_project_data["require-dbt-version"] = ">99999.0.0"
+        # allowed, because the RuntimeConfig checks, not the Project itself
+        project_from_config_norender(self.default_project_data, project_root=self.project_dir)
+
+    def test_none_values(self):
+        self.default_project_data.update(
+            {
+                "models": None,
+                "seeds": None,
+                "on-run-end": None,
+                "on-run-start": None,
+            }
+        )
+        project = project_from_config_rendered(self.default_project_data)
+        self.assertEqual(project.models, {})
+        self.assertEqual(project.on_run_start, [])
+        self.assertEqual(project.on_run_end, [])
+        self.assertEqual(project.seeds, {})
+
+    def test_nested_none_values(self):
+        self.default_project_data.update(
+            {
+                "models": {"vars": None, "pre-hook": None, "post-hook": None},
+                "seeds": {"vars": None, "pre-hook": None, "post-hook": None, "column_types": None},
+            }
+        )
+        project = project_from_config_rendered(self.default_project_data)
+        self.assertEqual(project.models, {"vars": {}, "pre-hook": [], "post-hook": []})
+        self.assertEqual(
+            project.seeds, {"vars": {}, "pre-hook": [], "post-hook": [], "column_types": {}}
+        )
+
+    @pytest.mark.skipif(os.name == "nt", reason="crashes CI for Windows")
+    def test_cycle(self):
+        models = {}
+        models["models"] = models
+        self.default_project_data.update(
+            {
+                "models": models,
+            }
+        )
+        with self.assertRaises(dbt.exceptions.DbtProjectError) as exc:
+            project_from_config_rendered(self.default_project_data)
+
+        assert "Cycle detected" in str(exc.exception)
+
+    def test_query_comment_disabled(self):
+        self.default_project_data.update(
+            {
+                "query-comment": None,
+            }
+        )
+        project = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir
+        )
+        self.assertEqual(project.query_comment.comment, "")
+        self.assertEqual(project.query_comment.append, False)
+
+        self.default_project_data.update(
+            {
+                "query-comment": "",
+            }
+        )
+        project = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir
+        )
+        self.assertEqual(project.query_comment.comment, "")
+        self.assertEqual(project.query_comment.append, False)
+
+    def test_default_query_comment(self):
+        project = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir
+        )
+        self.assertEqual(project.query_comment, QueryComment())
+
+    def test_default_query_comment_append(self):
+        self.default_project_data.update(
+            {
+                "query-comment": {"append": True},
+            }
+        )
+        project = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir
+        )
+        self.assertEqual(project.query_comment.comment, DEFAULT_QUERY_COMMENT)
+        self.assertEqual(project.query_comment.append, True)
+
+    def test_custom_query_comment_append(self):
+        self.default_project_data.update(
+            {
+                "query-comment": {"comment": "run by user test", "append": True},
+            }
+        )
+        project = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir
+        )
+        self.assertEqual(project.query_comment.comment, "run by user test")
+        self.assertEqual(project.query_comment.append, True)
+
+    def test_packages_from_dependencies(self):
+        packages = {
+            "packages": [
+                {
+                    "git": "{{ env_var('some_package') }}",
+                    "warn-unpinned": True,
+                }
+            ],
+        }
+
+        project = project_from_config_rendered(
+            self.default_project_data, packages, packages_specified_path=DEPENDENCIES_FILE_NAME
+        )
+        git_package = project.packages.packages[0]
+        # packages did not render because packages_specified_path=DEPENDENCIES_FILE_NAME
+        assert git_package.git == "{{ env_var('some_package') }}"
+
+
+class TestProjectFile(BaseConfigTest):
+    def test_from_project_root(self):
+        renderer = empty_project_renderer()
+        project = dbt.config.Project.from_project_root(self.project_dir, renderer)
+        from_config = project_from_config_norender(
+            self.default_project_data, project_root=self.project_dir
+        )
+        self.assertEqual(project, from_config)
+        self.assertEqual(project.version, "0.0.1")
+        self.assertEqual(project.project_name, "my_test_project")
+
+    def test_with_invalid_package(self):
+        renderer = empty_project_renderer()
+        self.write_packages({"invalid": ["not a package of any kind"]})
+        with self.assertRaises(dbt.exceptions.DbtProjectError):
+            dbt.config.Project.from_project_root(self.project_dir, renderer)
+
+
+class TestVariableProjectFile(BaseConfigTest):
+    def setUp(self):
+        super().setUp()
+        self.default_project_data["version"] = "{{ var('cli_version') }}"
+        self.default_project_data["name"] = "blah"
+        self.default_project_data["profile"] = "{{ env_var('env_value_profile') }}"
+        self.write_project(self.default_project_data)
+
+    def test_cli_and_env_vars(self):
+        renderer = dbt.config.renderer.DbtProjectYamlRenderer(None, {"cli_version": "0.1.2"})
+        with mock.patch.dict(os.environ, self.env_override):
+            safe_set_invocation_context()  # reset invocation context with new env
+            project = dbt.config.Project.from_project_root(
+                self.project_dir,
+                renderer,
+            )
+
+        self.assertEqual(renderer.ctx_obj.env_vars, {"env_value_profile": "default"})
+        self.assertEqual(project.version, "0.1.2")
+        self.assertEqual(project.project_name, "blah")
+        self.assertEqual(project.profile_name, "default")
+
+
+class TestVarLookups(unittest.TestCase):
+    def setUp(self):
+        self.initial_src_vars = {
+            # globals
+            "foo": 123,
+            "bar": "hello",
+            # project-scoped
+            "my_project": {
+                "bar": "goodbye",
+                "baz": True,
+            },
+            "other_project": {
+                "foo": 456,
+            },
+        }
+        self.src_vars = deepcopy(self.initial_src_vars)
+        self.dst = {"vars": deepcopy(self.initial_src_vars)}
+
+        self.projects = ["my_project", "other_project", "third_project"]
+        load_plugin("postgres")
+        self.local_var_search = mock.MagicMock(
+            fqn=["my_project", "my_model"], resource_type=NodeType.Model, package_name="my_project"
+        )
+        self.other_var_search = mock.MagicMock(
+            fqn=["other_project", "model"],
+            resource_type=NodeType.Model,
+            package_name="other_project",
+        )
+        self.third_var_search = mock.MagicMock(
+            fqn=["third_project", "third_model"],
+            resource_type=NodeType.Model,
+            package_name="third_project",
+        )
+
+    def test_lookups(self):
+        vars_provider = dbt.config.project.VarProvider(self.initial_src_vars)
+
+        expected = [
+            (self.local_var_search, "foo", 123),
+            (self.other_var_search, "foo", 456),
+            (self.third_var_search, "foo", 123),
+            (self.local_var_search, "bar", "goodbye"),
+            (self.other_var_search, "bar", "hello"),
+            (self.third_var_search, "bar", "hello"),
+            (self.local_var_search, "baz", True),
+            (self.other_var_search, "baz", None),
+            (self.third_var_search, "baz", None),
+        ]
+        for node, key, expected_value in expected:
+            value = vars_provider.vars_for(node, "postgres").get(key)
+            assert value == expected_value
+
+
+class TestMultipleProjectFlags(BaseConfigTest):
+    def setUp(self):
+        super().setUp()
+
+        self.default_project_data.update(
+            {
+                "flags": {
+                    "send_anonymous_usage_data": False,
+                }
+            }
+        )
+        self.write_project(self.default_project_data)
+
+        self.default_profile_data.update(
+            {
+                "config": {
+                    "send_anonymous_usage_data": False,
+                }
+            }
+        )
+        self.write_profile(self.default_profile_data)
+
+    def test_setting_multiple_flags(self):
+        with pytest.raises(dbt.exceptions.DbtProjectError):
+            set_from_args(self.args, None)
